% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{available_vocabularies}
\alias{available_vocabularies}
\title{Check Available Vocabularies}
\usage{
available_vocabularies(connection)
}
\arguments{
\item{connection}{a database connection to a valid OMOP database}
}
\value{
a list of availible vocabularies
}
\description{
Check Available Vocabularies
}
