% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vocab_constraints.R
\name{check_vocabulary_constraints}
\alias{check_vocabulary_constraints}
\title{Check Athena Vocabulary Constraints}
\usage{
check_vocabulary_constraints(
  vocabulary,
  path,
  action = c("remove_conflicts", "show_conflicts")
)
}
\arguments{
\item{vocabulary}{a list of vocabularies returned from \code{extract_vocab()}}

\item{path}{character string of the path to the location where you would like
to send the processed vocabularies}

\item{action}{character string of either: - "remove_conflicts": removes the
conflicting rows and writes out the data - "show_conflicts": keeps only the
conflicting rows and write out the data}
}
\value{
either \code{TRUE} or a list of vocabularies containing only violating
rows
}
\description{
Sometimes there are violations of the NULL and NOT NULL constants of the
athena vocabularies. If you are having trouble loading the vocabularies, this
might be why. This function will check the vocabularies that have been
imported by \code{extract_vocab()}. The function will write out new vocabularies
to the folder specified in \code{path} if taking \code{action = "remove_conflicts"}.
Otherwise if using \code{action = "show_conflicts"} only the rows that violate
constraints will be written out. In this case, the violations are also
returned in memory since there are likely to be a small number.
}
