% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{convert_to_standard}
\alias{convert_to_standard}
\title{Convert Non-Standard Vocabulary to Athena Standard}
\usage{
convert_to_standard(
  connection,
  concept_codes = diabetes$icd_10_code,
  source_vocabulary = "ICD10"
)
}
\arguments{
\item{connection}{a database connection to a valid OMOP database}

\item{concept_codes}{the concept codes you want to phenotype from the source
vocabulary specified in \code{source_vocabulary}}

\item{source_vocabulary}{character string of the source vocabulary you are
supplying}
}
\value{
a table with the standard representations for the non-standard codes
}
\description{
clinical concepts should be presented as "standard" concepts in OMOP. This
function takes non-standard codes, and returns their standard counterparts.
}
