% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{find_descendents}
\alias{find_descendents}
\title{Find Descendent Concepts}
\usage{
find_descendents(connection, concept_ids, depth)
}
\arguments{
\item{connection}{a database connection to a valid OMOP databse}

\item{concept_ids}{a list of standard athena ids.}

\item{depth}{integer vector (length 1) the number of descendent levels to
traverse}
}
\value{

}
\description{
Takes standard athena codes and returns any potential descendent terms (also
in standard form) government by the number of levels provided to \code{depth}
}
