% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vocab_constraints.R
\name{fix_vocabulary_constraints}
\alias{fix_vocabulary_constraints}
\title{Fix Athena Vocabulary Constraints}
\usage{
fix_vocabulary_constraints(vocab)
}
\arguments{
\item{vocab}{a list of vocabularies returned from \code{extract_vocab()}}
}
\value{
a vocab where null concept names are replaced with "No name provided"
}
\description{
Works like \code{check_vocabulary_constraints()} but fixes them.
}
