% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotype_athena}
\alias{phenotype_athena}
\title{Create an Athena based phenotype}
\usage{
phenotype_athena(
  connection,
  concept_codes = diabetes$icd_10_code,
  source_vocabulary = "ICD10",
  output_vocabularies = "everything"
)
}
\arguments{
\item{connection}{a database connection to a valid OMOP database}

\item{concept_codes}{the concept codes you want to phenotype from the source
vocabulary specified in \code{source_vocabulary}}

\item{source_vocabulary}{character vector (length 1) of the source vocabulary you are
supplying}

\item{output_vocabularies}{character vector (length n) listing the output
vocabularies for the phenotype. The default is "everything" which rather
unsurpringly returns everything}
}
\value{
a table with concept ids that relate to the input codes
}
\description{
Uses the Athena vocabularies to create a broad phenotype from any other
available vocabulary. The behaviour can be modified to limit the returning
vocabularies should it be desirable.
}
