% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype.R
\name{phenotype_athena}
\alias{phenotype_athena}
\title{Phenotype}
\usage{
phenotype_athena(
  connection,
  concept_codes = diabetes$icd_10_code,
  source_vocabulary = "ICD10"
)
}
\arguments{
\item{connection}{a database connection to a valid OMOP database}

\item{concept_codes}{the concept codes you want to phenotype from the source
vocabulary specified in \code{source_vocabulary}}

\item{source_vocabulary}{character string of the source vocabulary you are
supplying}
}
\value{

}
\description{
Phenotype
}
